Implement issue #3 - Using help messages in rustdoc.
The bulk of this is doing the new parsing needed in order
to be able to trim the content of help strings.
A normal person might use syn for this, but I'm avoiding
pulling in dependencies for relatively simple tasks I
can do myself and dodge noticeable compile time / dependency
download time hits on.